Add tls.Certificate to proxy struct Add ParseCa helper and improve certs.go
